feat: new post adapter created and used on home
